On FPGA Developer AMI use the correct vivado version
This is to enable having both vivado and sdx installed in the AMI
so can use both the master and prelease branches which use different
versions of vivado.

Use module load to use the correct vivado version based on the HDK version.

Save the vivado version used to build the models.

Rebuild the models if they were built with the wrong vivado version.

Add a lock file to prevent multiple process from trying to build the
models at the same time, potentially with different vivado versions.
This requires procmail to be installed otherwise the lockfile isn't created/used.

Merged changes from prelease.

Not downloading the shell dcp from S3.
Leaving that the same.
This shouldn't be an issue when switching between prelease and master
because they use different shell versions.