enhanced sdk install script for increased portability (#316)
* enhanced sdk install script for increased portability

Various Linux distributions have different preferred paths for installations of
unmanaged user binaries and libraries. This change attempts to follow those
preferences while maintaining maximum portability.

This script has been tested recent versions (at time of writing) of:
 - Amazon Linux 2017.03
 - Centos (FPGA Dev AMI 1.2)
 - Ubuntu 16.04.2 LTS
 - SUSE 12

It has not been tested on RHEL, but is presumed to work since it is similar
to Amazon Linux and Centos.

It installs binaries to /usr/local/bin if it's on the sudo PATH or /usr/bin
otherwise. It installs the shared library the first directory it finds searching
from the follwing list:
 /usr/local/lib64
 /usr/local/lib
 /usr/lib64
 /usr/lib
This usually results in installations to /usr/local/lib on Ubuntu and to
/usr/local/lib64 on others.

* removed "known bug" for Ubuntu