Synchronous load clear rescan (#29)
* Changing hdk_version to 1.2.1

* Fixing the programmer view

* spelling and grammer fixes

* enhanced sdk install script for increased portability

Various Linux distributions have different preferred paths for installations of
unmanaged user binaries and libraries. This change attempts to follow those
preferences while maintaining maximum portability.

This script has been tested recent versions (at time of writing) of:
 - Amazon Linux 2017.03
 - Centos (FPGA Dev AMI 1.2)
 - Ubuntu 16.04.2 LTS
 - SUSE 12

It has not been tested on RHEL, but is presumed to work since it is similar
to Amazon Linux and Centos.

It installs binaries to /usr/local/bin if it's on the sudo PATH or /usr/bin
otherwise. It installs the shared library the first directory it finds searching
from the follwing list:
 /usr/local/lib64
 /usr/local/lib
 /usr/lib64
 /usr/lib
This usually results in installations to /usr/local/lib on Ubuntu and to
/usr/local/lib64 on others.

* removed "known bug" for Ubuntu

* fixed an error return code

When the mailbox command timesout, return the -ETIMEDOUT. Negative
values are decoded from errno.h and positive number are defined by the
library.

* set the mbox timeout and delay when using --request-timeout

* updated fpga_pci_bar_attach to close fd after mmap

* Rj doc updates (#23)

* Placed the rescan option in bold text

* Placed the rescan option in bold text and italics

* Renaming Chipscope to Virtual JTAG; Additional clock recipes.

Change-Id: Ie4c8fb77893e011adcf85756c0e5d55e93f1fa9a

* Renaming disable chipscope define in sim .f files.

Change-Id: Ie99a86f1fd8639b614a618c6eb9a38a9e7358e4b

* Synchrnous load and clear operations with auto-rescan and end-state display (#27)

Synchronous load and clear operations with auto-rescan and end-state display

* Updated README.md (#28)