Add wait_for_afi.py
 ## Details
Waits for AFI creation to complete and optionally notify via email.

If --notify specified then it creates the SNS topic if it doesn't exist
and subscribes the email address to the topic if it isn't already subscribed.
If the subscription hasn't been verified it waits for it to be verified
so that the notification isn't lost.

Also modified notify_via_sns.py to allow SNS topic and email address to be
passed as parameters.
Also wait for subscription to be confirmed before doing the notification.

Moved shared SNS code to a shared python package in shared/lib.