HLx 1.4 shell fix for: invalid character "." (#420)
* HLx 1.4 shell fix for: invalid character "."

The piece of code I commented out was to support 2017.1 and older releases. Only 2017.1_sdx and newer releases are supported.
This code allowed 2017.1_sdxop, 2017.1_sdx, and later installs to work properly, but programmatically added the blocks when we were developing for 2017.1 (when it first came out before we added automation to the tools).
The issue came about because we used to call our tools 2017.1_op, now we call them 2017.4.op. The text split fails.

* fix for header

Fix for the header using github web again. If it still doesn't work I'll do it by command line.
Git hub wed does this: "We’ve detected the file encoding as ISO-8859-1. When you commit changes we will transcode it to UTF-8."
This must be the reason.

* Header fix using linux and not github Web...