Update path (#470)
* Update path

Attempting to change directories into this path (i.e., `/home/centos/src/project_data/aws-fpga/SDAccel/examples/xilinx_2017.4/getting_started/host/helloworld_ocl`) draws an error because the `xilinx_2017.4` directory is empty, now, and `xilinx` is a symbolic link that points to the latest release (i.e., `/home/centos/src/project_data/aws-fpga/SDAccel/examples/xilinx_2019.1`).  Additionally: `helloworld_ocl` is no longer a sub-directory of `host`; it's now a sub-directory of `hello_world`: so this part of the path should be updated as well.

* Fix unset flags

The alphanumeric part of these flags (i.e., `f`) are preceded by en dashes, when they should be hyphens; as using the former instead of the later results in errors (`-bash: unset: ``–f': not a valid identifier`).

* Update path to workspace directory

This change is consistent with 48c3048, and the basename of the image, but has a different filename extension because of the transparency.