Fixing Link To Vitis Examples (#494)
The old link to the Vitis examples in the README.md lead to https://github.com/aws/aws-fpga/blob/master/Vitis/examples/xilinx_2019.2, which is a 404 error. The new link is the result of going to https://github.com/aws/aws-fpga/blob/master/Vitis/examples and then entering the xilinx_2019.2 folder.